llm.py
from openai import OpenAI
import os
from dotenv import load_dotenv

load_dotenv()

client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))


def generate(context, question):
    prompt = f"""
    Use o contexto abaixo para responder à pergunta.

    Contexto:
    {context}

    Pergunta:
    {question}

    Responda de forma clara e objetiva.
    """

    response = client.chat.completions.create(
        model="gpt-4o-mini",
        messages=[
            {"role": "system", "content": "Você é um assistente útil."},
            {"role": "user", "content": prompt}
        ],
        temperature=0.3
    )

    return response.choices[0].message.content

