requirements.txt
fastapi==0.129.0
uvicorn==0.40.0

openai==2.21.0
python-dotenv==1.2.1

pymilvus==2.6.9
grpcio==1.78.0

boto3==1.42.49

pdfplumber==0.11.9
pdfminer.six==20251230
pypdfium2==5.4.0

numpy==2.4.2
pandas==3.0.0

tqdm==4.67.3